create modules for jbook 1 and jbook 2 toc parsing, add logic to determine if using myst.yaml or _toc.yaml and _config.yaml
